+++ packages/react/src/scene.ts
@@ -140,6 +140,7 @@ export type {
   SpotLightOptions,
   StandardMaterial,
   StandardMaterialOptions,
+  SurfaceDepth,
   StudioEnvironmentOptions,
   StudioEnvironmentLight,
   SolidTextureOptions,